% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_summary.R
\name{super_param_update}
\alias{super_param_update}
\title{super_param_update}
\usage{
super_param_update(
  Object = NULL,
  Path = NULL,
  name = NULL,
  strat_col = "start"
)
}
\arguments{
\item{Object}{\code{NULL}                           \code{NULL}}

\item{Path}{Excel           /}

\item{name}{Excel         \code{NULL}}

\item{strat_col}{\code{"start"}
#' @author      TjssnStat        Tongjissn;      support@https://study.tjbbb.com ;     Tongjissn     -         ;}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{                /       
}\if{html}{\out{</div>}}

\itemize{
\item \code{update$variable_diff}          /  /
\item \code{update$level_diff}               /
\item \code{update$RAW}
}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{#   1    Excel            (  Object            update  )
updated_obj <- super_param_update(
  Path = "data/2024_new_data.xlsx",  #     Excel  
  name = c("    ", "    "),  #        
  strat_col = "center"  #  "center"             
)
#            
head(updated_obj$update$RAW)

#   2                       super_param(          )
base_obj <- super_param_update(Path = "data/baseline.xlsx")  #     
final_obj <- super_param_update(
  Object = base_obj,  #         
  Path = "data/supplement.xlsx",  #       
  strat_col = "start"  #        
)
}\if{html}{\out{</div>}}
}
\description{
TjssnStat       \code{super_param_update()}
\code{excel_out(Object,Path,name)}
\enumerate{
\item      Excel
\item              /  /
\item
\item
}
}
\details{
Excel     TJssn_stat
\subsection{}{
\itemize{
\item **       **                                 90\%
\item **      **           \code{Object}
\item **         **    \code{strat_col}
\item **      **        \code{signal_step()}
}
}

\subsection{}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{/    `Object`               
}\if{html}{\out{</div>}}

\itemize{
\item \code{raw}
\item \code{dict}
\item \code{levels}
\item \code{update}                            \code{RAW}
}
\code{Object$update$RAW}
}
}
\keyword{Excel}
